chore: refactor to remove generic Fields and elements file
