+++ src/components/Form/ImageInput.field.tsx
@@ -1,14 +1,14 @@
 import * as React from 'react'
 import { Text } from 'theme-ui'
 import { ImageInput } from '../ImageInput/ImageInput'
-import { FieldContainer } from './elements'
-import type { IFieldProps } from './Fields'
+import { FieldContainer } from './FieldContainer'
+import type { FieldProps } from './types'
 
 // Assign correct typing so that ImageInput props can also be passed down to child
 // Note, partial as default onChange function provided
 type IImageInputProps = Partial<React.ComponentProps<typeof ImageInput>>
 
-interface IExtendedFieldProps extends IFieldProps, IImageInputProps {
+interface IExtendedFieldProps extends FieldProps, IImageInputProps {
   // add additional onChange style method to respond more directly to value changes
   // without need for react-final-form listener
   customChange?: (value) => void
