+++ src/components/Form/README.md
@@ -6,6 +6,5 @@ https://github.com/final-form/react-final-form
 It does not provide any of its own components, instead offering a `<Field />` wrapper that can
 take any existing component and hook into value and onChange props as required.
 
-The components in this folder are a combination of styled/wrapped standard input components
-(most of which can be found in the `Fields.tsx` file), and wrappers for custom components
+The components in this folder are a combination of styled/wrapped standard input components, and wrappers for custom components
 which sit outside of the main directory.
